Move SkipIntegrationTestInShortMode to testutil.
