+++ pkg/tests/utils.go
@@ -5,7 +5,6 @@ import (
 	"crypto/tls"
 	"net/url"
 	"os"
-	"strings"
 	"testing"
 
 	"github.com/go-openapi/strfmt"
@@ -24,16 +23,6 @@ import (
 	"github.com/grafana/grafana/pkg/services/user/userimpl"
 	"github.com/grafana/grafana/pkg/setting"
 )
-
-func SkipIntegrationTestInShortMode(t testing.TB) {
-	t.Helper()
-	if !strings.HasPrefix(t.Name(), "TestIntegration") {
-		t.Fatal("test is not an integration test")
-	}
-	if testing.Short() {
-		t.Skip("skipping integration test in short mode")
-	}
-}
 
 func CreateUser(t *testing.T, db db.DB, cfg *setting.Cfg, cmd user.CreateUserCommand) int64 {
 	t.Helper()
